+++ cpp/io/FileInputStream.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2007-2009 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2007-2010 Digital Bazaar, Inc. All rights reserved.
  */
 #ifndef monarch_io_FileInputStream_H
 #define monarch_io_FileInputStream_H
@@ -82,7 +82,7 @@ public:
     * @return the actual number of bytes skipped, or -1 if the end of the
     *         stream is reached or -1 if an IO exception occurred.
     */
-   virtual long long skip(long long count);
+   virtual int64_t skip(int64_t count);
 
    /**
     * Reads a line from the file.
